+++ common/src/main/java/club/iananderson/seasonhud/impl/seasons/mods/FabricSeasonsHelper.java
@@ -38,10 +38,10 @@ public class FabricSeasonsHelper {
     }
   }
 
-  public static int getDate(Player player) {
-    int dayLength = Config.getDayLength();
-    int seasonLength = FabricSeasons.CONFIG.getSpringLength();
-    int worldTime = Math.toIntExact(player.level().getDayTime());
+  public static long getDate(Player player) {
+    long dayLength = Config.getDayLength();
+    long seasonLength = FabricSeasons.CONFIG.getSpringLength();
+    long worldTime = Math.toIntExact(player.level().getDayTime());
 
     // Get the current day of month from the system. Used with fabric seasons' system time tied with season option
     if (FabricSeasonsHelper.isSeasonTiedWithSystemTime) {
